Move StatsManager.php to App\Slave namespace
