+++ src/Slave/Worker/StatsManager.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\DependencyInjection;
+namespace App\Slave\Worker;
 
 use App\Slave\Task\FetchConfiguration;
 use App\Slave\Task\PublishResults;
